Leaf 4042 - moved the file check

+++ LEAF_Request_Portal/mailer/mailer.php
@@ -56,15 +56,13 @@ foreach ($queue as $item)
     if (strlen($item) == 40)
     {
         // attempt to resend email if its 5 minutes old
-        if (time() - filemtime($folder . $item) >= 300)
+        if (file_exists($folder . $item) && time() - filemtime($folder . $item) >= 300)
         {
             $email = unserialize(file_get_contents($folder . $item));
             if (strlen(trim($email['recipient'])) == 0)
             {
                 // delete invalid cache
-                if (file_exists($folder . $item)) {
-                    unlink($folder . $item);
-                }
+                unlink($folder . $item);
 
                 trigger_error('Mail no recipient: ' . $email['subject']);
             }
@@ -73,9 +71,7 @@ foreach ($queue as $item)
                 touch($folder . $item);    // reset timer
                 if (mail($email['recipient'], $email['subject'], $email['body'], $email['headers']))
                 {
-                    if (file_exists($folder . $item)) {
-                        unlink($folder . $item);
-                    }
+                    unlink($folder . $item);
                     trigger_error('Queued mail sent: ' . $email['subject']);
                     if ($webMode)
                     {
